+++ src/main/java/com/desafiopicpay/picpay/domain/transaction/TransactionRequestDTO.java
@@ -1,6 +1,7 @@
 package com.desafiopicpay.picpay.domain.transaction;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record TransactionRequestDTO(UUID payerID, UUID payeeID, Long value) {
+public record TransactionRequestDTO(UUID payerID, UUID payeeID, BigDecimal value) {
 }
